fix and update public health controller and dashboard tests
