+++ test/controllers/jurisdictions_controller_test.rb
@@ -31,8 +31,8 @@ class JurisdictionsControllerTest < ActionController::TestCase
 
         assert_equal user_jur.subtree.size, json_response['jurisdiction_paths'].size
         user_jur.subtree.each do |sub_jur|
-          assert json_response['jurisdiction_paths'].key?(sub_jur[:id].to_s)
-          assert_equal sub_jur[:path], json_response['jurisdiction_paths'][sub_jur[:id].to_s]
+          assert json_response['jurisdiction_paths'].key?(sub_jur.id.to_s)
+          assert_equal sub_jur[:path], json_response['jurisdiction_paths'][sub_jur.id.to_s]
         end
 
         sign_out user
@@ -41,13 +41,13 @@ class JurisdictionsControllerTest < ActionController::TestCase
   end
 
   test 'assigned users for viewable patients' do
-    get :assigned_users_for_viewable_patients
+    post :assigned_users_for_viewable_patients
     assert_redirected_to(new_user_session_path)
 
     %i[analyst_user].each do |role|
       user = create(role)
       sign_in user
-      get :jurisdiction_paths
+      post :jurisdiction_paths
       assert_response :unauthorized
       sign_out user
     end
@@ -61,93 +61,125 @@ class JurisdictionsControllerTest < ActionController::TestCase
         sign_in user
 
         error = assert_raises(ActionController::ParameterMissing) do
-          get :assigned_users_for_viewable_patients
+          post :assigned_users_for_viewable_patients
         end
-        assert_includes(error.message, 'jurisdiction_id')
+        assert_includes(error.message, 'query')
 
         error = assert_raises(ActionController::ParameterMissing) do
-          get :assigned_users_for_viewable_patients, params: { jurisdiction_id: 4 }
+          post :assigned_users_for_viewable_patients, params: { query: { scope: '' } }, as: :json
+        end
+        assert_includes(error.message, 'jurisdiction')
+
+        error = assert_raises(ActionController::ParameterMissing) do
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: 4 } }, as: :json
         end
         assert_includes(error.message, 'scope')
 
-        get :assigned_users_for_viewable_patients, params: { jurisdiction_id: 4, scope: 'all', workflow: 'asdf' }
+        post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: 4, scope: 'all', workflow: 'asdf' } }, as: :json
         assert_response :bad_request
 
-        get :assigned_users_for_viewable_patients, params: { jurisdiction_id: 4, scope: 'all', tab: 'all' }
+        post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: 4, scope: 'all', tab: 'fdsa' } }, as: :json
         assert_response :bad_request
 
-        get :assigned_users_for_viewable_patients, params: { jurisdiction_id: 4, scope: 'all', workflow: 'exposure', tab: 'requiring_review' }
+        post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: 4, scope: 'all', workflow: 'exposure', tab: 'reporting' } }, as: :json
         assert_response :bad_request
 
-        get :assigned_users_for_viewable_patients, params: { jurisdiction_id: 4, scope: 'all', workflow: 'isolation', tab: 'pui' }
+        post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: 4, scope: 'all', workflow: 'isolation', tab: 'pui' } }, as: :json
         assert_response :bad_request
 
         sign_out user
       end
     end
 
-    user = create(:public_health_user)
-
-    jurs = Jurisdiction.where(path: ['USA', 'USA, State 1', 'USA, State 1, County 1'])
-    jurs.find_each do |user_jur|
-      user.update(jurisdiction: user_jur)
+    Jurisdiction.where(path: ['USA', 'USA, State 1', 'USA, State 1, County 1']).find_each do |user_jur|
+      user = create(:public_health_user, jurisdiction: user_jur)
       sign_in user
 
-      jurs.find_each do |jur|
+      Jurisdiction.all.find_each do |jur|
         %w[all exact].each do |scope|
-          get :assigned_users_for_viewable_patients, params: { jurisdiction_id: jur[:id], scope: scope }
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope } }, as: :json
           assert_response :bad_request && next unless user_jur.subtree.include?(jur)
 
-          patients = scope == 'all' ? jur.all_patients.where.not(assigned_user: nil) : jur.immediate_patients.where.not(assigned_user: nil)
+          patients = scope == 'exact' ? jur.immediate_patients.where.not(assigned_user: nil) : jur.all_patients.where.not(assigned_user: nil)
           assert_equal patients.distinct.pluck(:assigned_user).sort, JSON.parse(response.body)['assigned_users']
 
-          get :assigned_users_for_viewable_patients, params: { jurisdiction_id: jur[:id], scope: scope, workflow: 'exposure', tab: 'all' }
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'exposure', tab: 'all' } },
+                                                      as: :json
           assert_equal patients.where(isolation: false, purged: false).distinct.pluck(:assigned_user).sort, JSON.parse(response.body)['assigned_users']
 
-          get :assigned_users_for_viewable_patients, params: { jurisdiction_id: jur[:id], scope: scope, workflow: 'exposure', tab: 'symptomatic' }
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'exposure', tab: 'symptomatic' } },
+                                                      as: :json
           assert_equal patients.exposure_symptomatic.pluck(:assigned_user).sort, JSON.parse(response.body)['assigned_users']
 
-          get :assigned_users_for_viewable_patients, params: { jurisdiction_id: jur[:id], scope: scope, workflow: 'exposure', tab: 'non_reporting' }
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'exposure', tab: 'non_reporting' } },
+                                                      as: :json
           assert_equal patients.exposure_non_reporting.pluck(:assigned_user).sort, JSON.parse(response.body)['assigned_users']
 
-          get :assigned_users_for_viewable_patients, params: { jurisdiction_id: jur[:id], scope: scope, workflow: 'exposure', tab: 'asymptomatic' }
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'exposure', tab: 'asymptomatic' } },
+                                                      as: :json
           assert_equal patients.exposure_asymptomatic.pluck(:assigned_user).sort, JSON.parse(response.body)['assigned_users']
 
-          get :assigned_users_for_viewable_patients, params: { jurisdiction_id: jur[:id], scope: scope, workflow: 'exposure', tab: 'pui' }
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'exposure', tab: 'pui' } },
+                                                      as: :json
           assert_equal patients.exposure_under_investigation.pluck(:assigned_user).sort, JSON.parse(response.body)['assigned_users']
 
-          get :assigned_users_for_viewable_patients, params: { jurisdiction_id: jur[:id], scope: scope, workflow: 'exposure', tab: 'closed' }
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'exposure', tab: 'closed' } },
+                                                      as: :json
           assigned_users = patients.where(isolation: false, monitoring: false, purged: false).pluck(:assigned_user).sort
           assert_equal assigned_users, JSON.parse(response.body)['assigned_users']
 
-          get :assigned_users_for_viewable_patients, params: { jurisdiction_id: jur[:id], scope: scope, workflow: 'exposure', tab: 'transferred_in' }
-          assigned_users = if scope == 'all'
-                             jur.transferred_in_patients.where(isolation: false).pluck(:assigned_user).sort
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'exposure', tab: 'transferred_in' } },
+                                                      as: :json
+          assigned_users = if scope == 'exact'
+                             jur.transferred_in_patients.where(isolation: false, jurisdiction: jur.id).where.not(assigned_user: nil).pluck(:assigned_user).sort
                            else
-                             jur.transferred_in_patients.where(isolation: false, jurisdiction_id: jur[:id]).pluck(:assigned_user).sort
+                             jur.transferred_in_patients.where(isolation: false).where.not(assigned_user: nil).pluck(:assigned_user).sort
                            end
           assert_equal assigned_users, JSON.parse(response.body)['assigned_users']
 
-          get :assigned_users_for_viewable_patients, params: { jurisdiction_id: jur[:id], scope: scope, workflow: 'isolation', tab: 'all' }
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'isolation', tab: 'all' } },
+                                                      as: :json
           assert_equal patients.where(isolation: true, purged: false).distinct.pluck(:assigned_user).sort, JSON.parse(response.body)['assigned_users']
 
-          get :assigned_users_for_viewable_patients, params: { jurisdiction_id: jur[:id], scope: scope, workflow: 'isolation', tab: 'requiring_review' }
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'isolation', tab: 'requiring_review' } },
+                                                      as: :json
           assert_equal patients.isolation_requiring_review.pluck(:assigned_user).sort, JSON.parse(response.body)['assigned_users']
 
-          get :assigned_users_for_viewable_patients, params: { jurisdiction_id: jur[:id], scope: scope, workflow: 'isolation', tab: 'non_reporting' }
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'isolation', tab: 'non_reporting' } },
+                                                      as: :json
           assert_equal patients.isolation_non_reporting.pluck(:assigned_user).sort, JSON.parse(response.body)['assigned_users']
 
-          get :assigned_users_for_viewable_patients, params: { jurisdiction_id: jur[:id], scope: scope, workflow: 'isolation', tab: 'reporting' }
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'isolation', tab: 'reporting' } },
+                                                      as: :json
           assert_equal patients.isolation_reporting.pluck(:assigned_user).sort, JSON.parse(response.body)['assigned_users']
 
-          get :assigned_users_for_viewable_patients, params: { jurisdiction_id: jur[:id], scope: scope, workflow: 'isolation', tab: 'closed' }
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'isolation', tab: 'closed' } },
+                                                      as: :json
           assert_equal patients.where(isolation: true, monitoring: false, purged: false).pluck(:assigned_user).sort, JSON.parse(response.body)['assigned_users']
 
-          get :assigned_users_for_viewable_patients, params: { jurisdiction_id: jur[:id], scope: scope, workflow: 'isolation', tab: 'transferred_in' }
-          assigned_users = if scope == 'all'
-                             jur.transferred_in_patients.where(isolation: true).pluck(:assigned_user).sort
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'isolation', tab: 'transferred_in' } },
+                                                      as: :json
+          assigned_users = if scope == 'exact'
+                             jur.transferred_in_patients.where(isolation: true, jurisdiction: jur.id).where.not(assigned_user: nil).pluck(:assigned_user).sort
                            else
-                             jur.transferred_in_patients.where(isolation: true, jurisdiction_id: jur[:id]).pluck(:assigned_user).sort
+                             jur.transferred_in_patients.where(isolation: true).where.not(assigned_user: nil).pluck(:assigned_user).sort
+                           end
+          assert_equal assigned_users, JSON.parse(response.body)['assigned_users']
+
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'all', tab: 'all' } },
+                                                      as: :json
+          assert_equal patients.where(purged: false).distinct.pluck(:assigned_user).sort, JSON.parse(response.body)['assigned_users']
+
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'all', tab: 'closed' } },
+                                                      as: :json
+          assert_equal patients.where(monitoring: false, purged: false).pluck(:assigned_user).sort, JSON.parse(response.body)['assigned_users']
+
+          post :assigned_users_for_viewable_patients, params: { query: { jurisdiction: jur.id, scope: scope, workflow: 'all', tab: 'transferred_in' } },
+                                                      as: :json
+          assigned_users = if scope == 'exact'
+                             jur.transferred_in_patients.where(jurisdiction: jur.id).where.not(assigned_user: nil).pluck(:assigned_user).sort
+                           else
+                             jur.transferred_in_patients.where.not(assigned_user: nil).pluck(:assigned_user).sort
                            end
           assert_equal assigned_users, JSON.parse(response.body)['assigned_users']
         end
