+++ test/controllers/public_health_controller_test.rb
@@ -8,13 +8,13 @@ class PublicHealthControllerTest < ActionController::TestCase
   def teardown; end
 
   test 'patients authorization' do
-    get :patients
+    post :patients
     assert_redirected_to(new_user_session_path)
 
     %i[admin_user analyst_user enroller_user].each do |role|
       user = create(role)
       sign_in user
-      get :patients
+      post :patients
       assert_redirected_to @controller.root_url
       sign_out user
     end
@@ -25,12 +25,17 @@ class PublicHealthControllerTest < ActionController::TestCase
         sign_in user
 
         error = assert_raises(ActionController::ParameterMissing) do
-          get :patients
+          post :patients
+        end
+        assert_includes(error.message, 'query')
+
+        error = assert_raises(ActionController::ParameterMissing) do
+          post :patients, params: { query: { tab: 'all' } }, as: :json
         end
         assert_includes(error.message, 'workflow')
 
         error = assert_raises(ActionController::ParameterMissing) do
-          get :patients, params: { workflow: 'exposure' }
+          post :patients, params: { query: { workflow: 'exposure' } }, as: :json
         end
         assert_includes(error.message, 'tab')
 
@@ -40,71 +45,71 @@ class PublicHealthControllerTest < ActionController::TestCase
   end
 
   test 'patients param validation' do
-    user = create(:public_health_user, jurisdiction: Jurisdiction.where(path: 'USA, State 1').first)
+    user = create(:public_health_user, jurisdiction: Jurisdiction.find_by(path: 'USA, State 1'))
     sign_in user
 
-    get :patients, params: { workflow: 'asdf', tab: 'all' }
+    post :patients, params: { query: { workflow: 'asdf', tab: 'all' } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'exposure', tab: 'reporting' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'reporting' } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'isolation', tab: 'pui' }
+    post :patients, params: { query: { workflow: 'isolation', tab: 'pui' } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'exposure', tab: 'symptomatic', jurisdiction: 'asdf' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'symptomatic', jurisdiction: 'asdf' } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'exposure', tab: 'symptomatic', jurisdiction: Jurisdiction.where(path: 'USA, State 2').first }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'symptomatic', jurisdiction: Jurisdiction.find_by(path: 'USA, State 2').id } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'exposure', tab: 'symptomatic', scope: 'fdsa' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'symptomatic', scope: 'fdsa' } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'exposure', tab: 'symptomatic', user: 'asdf' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'symptomatic', user: 'asdf' } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'exposure', tab: 'symptomatic', user: '0' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'symptomatic', user: '0' } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'exposure', tab: 'symptomatic', user: '10000' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'symptomatic', user: '10000' } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'exposure', tab: 'symptomatic', entries: '-1' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'symptomatic', entries: '-1' } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'exposure', tab: 'symptomatic', page: '-1' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'symptomatic', page: '-1' } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'exposure', tab: 'symptomatic', order: 'asdf' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'symptomatic', order: 'asdf' } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'exposure', tab: 'symptomatic', direction: 'fdsa' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'symptomatic', direction: 'fdsa' } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'exposure', tab: 'symptomatic', order: 'name' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'symptomatic', order: 'name' } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'exposure', tab: 'symptomatic', direction: 'asc' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'symptomatic', direction: 'asc' } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'exposure', tab: 'symptomatic', order: '', direction: 'asc' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'symptomatic', order: '', direction: 'asc' } }, as: :json
     assert_response :bad_request
 
-    get :patients, params: { workflow: 'exposure', tab: 'symptomatic', order: 'name', direction: '' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'symptomatic', order: 'name', direction: '' } }, as: :json
     assert_response :bad_request
 
     sign_out user
   end
 
-  test 'patients by workflow and tab' do
+  test 'patients by linelist' do
     Jurisdiction.where(path: ['USA, State 1', 'USA, State 1, County 1']).find_each do |user_jur|
       user = create(:public_health_user, jurisdiction: user_jur)
       sign_in user
 
       common_fields = %w[name state_local_id dob]
 
-      get :patients, params: { workflow: 'exposure', tab: 'symptomatic' }
+      post :patients, params: { query: { workflow: 'exposure', tab: 'symptomatic' } }, as: :json
       json_response = JSON.parse(response.body)
       patients = user.viewable_patients.exposure_symptomatic
       assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
@@ -112,7 +117,7 @@ class PublicHealthControllerTest < ActionController::TestCase
       assert_equal common_fields + %w[jurisdiction assigned_user end_of_monitoring risk_level
                                       monitoring_plan latest_report report_eligibility], json_response['fields']
 
-      get :patients, params: { workflow: 'exposure', tab: 'non_reporting' }
+      post :patients, params: { query: { workflow: 'exposure', tab: 'non_reporting' } }, as: :json
       json_response = JSON.parse(response.body)
       patients = user.viewable_patients.exposure_non_reporting
       assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
@@ -120,7 +125,7 @@ class PublicHealthControllerTest < ActionController::TestCase
       assert_equal common_fields + %w[jurisdiction assigned_user end_of_monitoring risk_level
                                       monitoring_plan latest_report report_eligibility], json_response['fields']
 
-      get :patients, params: { workflow: 'exposure', tab: 'asymptomatic' }
+      post :patients, params: { query: { workflow: 'exposure', tab: 'asymptomatic' } }, as: :json
       json_response = JSON.parse(response.body)
       patients = user.viewable_patients.exposure_asymptomatic
       assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
@@ -128,7 +133,7 @@ class PublicHealthControllerTest < ActionController::TestCase
       assert_equal common_fields + %w[jurisdiction assigned_user end_of_monitoring risk_level
                                       monitoring_plan latest_report report_eligibility], json_response['fields']
 
-      get :patients, params: { workflow: 'exposure', tab: 'pui' }
+      post :patients, params: { query: { workflow: 'exposure', tab: 'pui' } }, as: :json
       json_response = JSON.parse(response.body)
       patients = user.viewable_patients.exposure_under_investigation
       assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
@@ -136,28 +141,28 @@ class PublicHealthControllerTest < ActionController::TestCase
       assert_equal common_fields + %w[jurisdiction assigned_user end_of_monitoring risk_level
                                       public_health_action latest_report report_eligibility], json_response['fields']
 
-      get :patients, params: { workflow: 'exposure', tab: 'closed' }
+      post :patients, params: { query: { workflow: 'exposure', tab: 'closed' } }, as: :json
       json_response = JSON.parse(response.body)
       patients = user.viewable_patients.monitoring_closed_without_purged.where(isolation: false)
       assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
       assert_equal patients.size, json_response['total']
       assert_equal common_fields + %w[jurisdiction assigned_user expected_purge_date reason_for_closure closed_at], json_response['fields']
 
-      get :patients, params: { workflow: 'exposure', tab: 'transferred_in' }
+      post :patients, params: { query: { workflow: 'exposure', tab: 'transferred_in' } }, as: :json
       json_response = JSON.parse(response.body)
-      patients = user.jurisdiction.transferred_in_patients.where(isolation: false)
+      patients = user_jur.transferred_in_patients.where(isolation: false)
       assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
       assert_equal patients.size, json_response['total']
       assert_equal common_fields + %w[transferred_from end_of_monitoring risk_level monitoring_plan transferred_at], json_response['fields']
 
-      get :patients, params: { workflow: 'exposure', tab: 'transferred_out' }
+      post :patients, params: { query: { workflow: 'exposure', tab: 'transferred_out' } }, as: :json
       json_response = JSON.parse(response.body)
-      patients = user.jurisdiction.transferred_out_patients.where(isolation: false)
+      patients = user_jur.transferred_out_patients.where(isolation: false)
       assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
       assert_equal patients.size, json_response['total']
       assert_equal common_fields + %w[transferred_to end_of_monitoring risk_level monitoring_plan transferred_at], json_response['fields']
 
-      get :patients, params: { workflow: 'exposure', tab: 'all' }
+      post :patients, params: { query: { workflow: 'exposure', tab: 'all' } }, as: :json
       json_response = JSON.parse(response.body)
       patients = user.viewable_patients.where(isolation: false, purged: false)
       assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
@@ -165,7 +170,7 @@ class PublicHealthControllerTest < ActionController::TestCase
       assert_equal common_fields + %w[jurisdiction assigned_user end_of_monitoring risk_level
                                       monitoring_plan latest_report status report_eligibility], json_response['fields']
 
-      get :patients, params: { workflow: 'isolation', tab: 'requiring_review' }
+      post :patients, params: { query: { workflow: 'isolation', tab: 'requiring_review' } }, as: :json
       json_response = JSON.parse(response.body)
       patients = user.viewable_patients.isolation_requiring_review
       assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
@@ -173,7 +178,7 @@ class PublicHealthControllerTest < ActionController::TestCase
       assert_equal common_fields + %w[jurisdiction assigned_user extended_isolation symptom_onset monitoring_plan
                                       latest_report report_eligibility], json_response['fields']
 
-      get :patients, params: { workflow: 'isolation', tab: 'non_reporting' }
+      post :patients, params: { query: { workflow: 'isolation', tab: 'non_reporting' } }, as: :json
       json_response = JSON.parse(response.body)
       patients = user.viewable_patients.isolation_non_reporting
       assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
@@ -181,7 +186,7 @@ class PublicHealthControllerTest < ActionController::TestCase
       assert_equal common_fields + %w[jurisdiction assigned_user extended_isolation symptom_onset monitoring_plan
                                       latest_report report_eligibility], json_response['fields']
 
-      get :patients, params: { workflow: 'isolation', tab: 'reporting' }
+      post :patients, params: { query: { workflow: 'isolation', tab: 'reporting' } }, as: :json
       json_response = JSON.parse(response.body)
       patients = user.viewable_patients.isolation_reporting
       assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
@@ -189,28 +194,28 @@ class PublicHealthControllerTest < ActionController::TestCase
       assert_equal common_fields + %w[jurisdiction assigned_user extended_isolation symptom_onset monitoring_plan
                                       latest_report report_eligibility], json_response['fields']
 
-      get :patients, params: { workflow: 'isolation', tab: 'closed' }
+      post :patients, params: { query: { workflow: 'isolation', tab: 'closed' } }, as: :json
       json_response = JSON.parse(response.body)
       patients = user.viewable_patients.monitoring_closed_without_purged.where(isolation: true)
       assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
       assert_equal patients.size, json_response['total']
       assert_equal common_fields + %w[jurisdiction assigned_user expected_purge_date reason_for_closure closed_at], json_response['fields']
 
-      get :patients, params: { workflow: 'isolation', tab: 'transferred_in' }
+      post :patients, params: { query: { workflow: 'isolation', tab: 'transferred_in' } }, as: :json
       json_response = JSON.parse(response.body)
-      patients = user.jurisdiction.transferred_in_patients.where(isolation: true)
+      patients = user_jur.transferred_in_patients.where(isolation: true)
       assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
       assert_equal patients.size, json_response['total']
       assert_equal common_fields + %w[transferred_from monitoring_plan transferred_at], json_response['fields']
 
-      get :patients, params: { workflow: 'isolation', tab: 'transferred_out' }
+      post :patients, params: { query: { workflow: 'isolation', tab: 'transferred_out' } }, as: :json
       json_response = JSON.parse(response.body)
-      patients = user.jurisdiction.transferred_out_patients.where(isolation: true)
+      patients = user_jur.transferred_out_patients.where(isolation: true)
       assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
       assert_equal patients.size, json_response['total']
       assert_equal common_fields + %w[transferred_to monitoring_plan transferred_at], json_response['fields']
 
-      get :patients, params: { workflow: 'isolation', tab: 'all' }
+      post :patients, params: { query: { workflow: 'isolation', tab: 'all' } }, as: :json
       json_response = JSON.parse(response.body)
       patients = user.viewable_patients.where(isolation: true, purged: false)
       assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
@@ -218,57 +223,81 @@ class PublicHealthControllerTest < ActionController::TestCase
       assert_equal common_fields + %w[jurisdiction assigned_user extended_isolation symptom_onset monitoring_plan
                                       latest_report status report_eligibility], json_response['fields']
 
+      post :patients, params: { query: { workflow: 'all', tab: 'closed' } }, as: :json
+      json_response = JSON.parse(response.body)
+      patients = user.viewable_patients.monitoring_closed_without_purged
+      assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
+      assert_equal patients.size, json_response['total']
+
+      post :patients, params: { query: { workflow: 'all', tab: 'transferred_in' } }, as: :json
+      json_response = JSON.parse(response.body)
+      patients = user_jur.transferred_in_patients
+      assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
+      assert_equal patients.size, json_response['total']
+
+      post :patients, params: { query: { workflow: 'all', tab: 'transferred_out' } }, as: :json
+      json_response = JSON.parse(response.body)
+      patients = user_jur.transferred_out_patients
+      assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
+      assert_equal patients.size, json_response['total']
+
+      post :patients, params: { query: { workflow: 'all', tab: 'all', entries: 100 } }, as: :json
+      json_response = JSON.parse(response.body)
+      patients = user.viewable_patients.where(purged: false)
+      assert_equal patients.order(:id).pluck(:id), json_response['linelist'].map { |patient| patient['id'] }.sort
+      assert_equal patients.size, json_response['total']
+
       sign_out user
     end
   end
 
   test 'patients by jurisdiction and assigned user' do
-    user = create(:public_health_user, jurisdiction: Jurisdiction.where(path: 'USA, State 1').first)
+    user = create(:public_health_user, jurisdiction: Jurisdiction.find_by(path: 'USA, State 1'))
     sign_in user
 
-    jur = Jurisdiction.where(path: 'USA, State 1, County 1').first
+    jur = Jurisdiction.find_by(path: 'USA, State 1, County 1')
 
-    get :patients, params: { workflow: 'exposure', tab: 'all', jurisdiction: jur[:id], scope: 'all' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'all', jurisdiction: jur[:id], scope: 'all' } }, as: :json
     JSON.parse(response.body)['linelist'].each do |patient|
       assert jur.subtree.pluck(:name).include?(patient['jurisdiction'])
     end
 
-    get :patients, params: { workflow: 'exposure', tab: 'all', jurisdiction: jur[:id], scope: 'exact' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'all', jurisdiction: jur[:id], scope: 'exact' } }, as: :json
     JSON.parse(response.body)['linelist'].each do |patient|
       assert_equal jur[:name], patient['jurisdiction']
     end
 
-    get :patients, params: { workflow: 'exposure', tab: 'all', user: 'none' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'all', user: 'none' } }, as: :json
     JSON.parse(response.body)['linelist'].each do |patient|
       assert patient['assigned_user'].blank?
     end
 
     assigned_user = user.viewable_patients.where.not(assigned_user: nil).distinct.pluck(:assigned_user).first
 
-    get :patients, params: { workflow: 'exposure', tab: 'all', user: assigned_user }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'all', user: assigned_user } }, as: :json
     JSON.parse(response.body)['linelist'].each do |patient|
       assert_equal assigned_user, patient['assigned_user']
     end
 
-    get :patients, params: { workflow: 'exposure', tab: 'all', jurisdiction: jur[:id], scope: 'all', user: 'none' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'all', jurisdiction: jur[:id], scope: 'all', user: 'none' } }, as: :json
     JSON.parse(response.body)['linelist'].each do |patient|
       assert jur.subtree.pluck(:name).include?(patient['jurisdiction'])
       assert patient['assigned_user'].blank?
     end
 
-    get :patients, params: { workflow: 'exposure', tab: 'all', jurisdiction: jur[:id], scope: 'exact', user: 'none' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'all', jurisdiction: jur[:id], scope: 'exact', user: 'none' } }, as: :json
     JSON.parse(response.body)['linelist'].each do |patient|
       assert_equal jur[:name], patient['jurisdiction']
       assert patient['assigned_user'].blank?
     end
 
-    get :patients, params: { workflow: 'exposure', tab: 'all', jurisdiction: jur[:id], scope: 'all', user: assigned_user }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'all', jurisdiction: jur[:id], scope: 'all', user: assigned_user } }, as: :json
     JSON.parse(response.body)['linelist'].each do |patient|
       assert jur.subtree.pluck(:name).include?(patient['jurisdiction'])
       assert_equal assigned_user, patient['assigned_user']
     end
 
-    get :patients, params: { workflow: 'exposure', tab: 'all', jurisdiction: jur[:id], scope: 'exact', user: assigned_user }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'all', jurisdiction: jur[:id], scope: 'exact', user: assigned_user } }, as: :json
     JSON.parse(response.body)['linelist'].each do |patient|
       assert_equal jur[:name], patient['jurisdiction']
       assert_equal assigned_user, patient['assigned_user']
@@ -278,11 +307,11 @@ class PublicHealthControllerTest < ActionController::TestCase
   end
 
   test 'patients filtering' do
-    user = create(:public_health_user, jurisdiction: Jurisdiction.where(path: 'USA, State 1').first)
+    user = create(:public_health_user, jurisdiction: Jurisdiction.find_by(path: 'USA, State 1'))
     sign_in user
 
     filtered_patient = user.viewable_patients.where(isolation: false, purged: false).first
-    get :patients, params: { workflow: 'exposure', tab: 'all', search: filtered_patient[:first_name] }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'all', search: filtered_patient[:first_name] } }, as: :json
     JSON.parse(response.body)['linelist'].each do |patient|
       assert patient['name'].include?(filtered_patient[:first_name])
     end
@@ -291,16 +320,16 @@ class PublicHealthControllerTest < ActionController::TestCase
   end
 
   test 'patients sorting' do
-    user = create(:public_health_user, jurisdiction: Jurisdiction.where(path: 'USA, State 1').first)
+    user = create(:public_health_user, jurisdiction: Jurisdiction.find_by(path: 'USA, State 1'))
     sign_in user
 
     order = 'name'
 
-    get :patients, params: { workflow: 'exposure', tab: 'all', order: order, direction: 'asc' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'all', order: order, direction: 'asc' } }, as: :json
     patient_ids = user.viewable_patients.where(isolation: false, purged: false).order(:last_name, :first_name).pluck(:id)
     assert_equal patient_ids, (JSON.parse(response.body)['linelist'].map { |patient| patient['id'] })
 
-    get :patients, params: { workflow: 'exposure', tab: 'all', order: order, direction: 'desc' }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'all', order: order, direction: 'desc' } }, as: :json
     patient_ids = user.viewable_patients.where(isolation: false, purged: false).order(last_name: :desc, first_name: :desc).pluck(:id)
     assert_equal patient_ids, (JSON.parse(response.body)['linelist'].map { |patient| patient['id'] })
 
@@ -308,17 +337,17 @@ class PublicHealthControllerTest < ActionController::TestCase
   end
 
   test 'patients pagination' do
-    user = create(:public_health_user, jurisdiction: Jurisdiction.where(path: 'USA, State 1').first)
+    user = create(:public_health_user, jurisdiction: Jurisdiction.find_by(path: 'USA, State 1'))
     sign_in user
 
     entries = 5
 
-    get :patients, params: { workflow: 'exposure', tab: 'all', entries: entries, page: 0 }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'all', entries: entries, page: 0 } }, as: :json
     page_0 = JSON.parse(response.body)
     assert_equal entries, page_0['linelist'].size
     assert_not_equal entries, page_0['total']
 
-    get :patients, params: { workflow: 'exposure', tab: 'all', entries: entries, page: 1 }
+    post :patients, params: { query: { workflow: 'exposure', tab: 'all', entries: entries, page: 1 } }, as: :json
     page_1 = JSON.parse(response.body)
     assert_not_equal page_0, page_1
 
@@ -351,6 +380,16 @@ class PublicHealthControllerTest < ActionController::TestCase
         sign_out user
       end
     end
+  end
+
+  test 'patient counts' do
+    user = create(:public_health_user, jurisdiction: Jurisdiction.find_by(path: 'USA'))
+    sign_in user
+
+    post :patients_count, params: { query: { workflow: 'all', tab: 'all', jurisdiction: user.jurisdiction.id } }, as: :json
+    assert_equal user.jurisdiction.all_patients.size, JSON.parse(response.body)['count']
+
+    sign_out user
   end
 
   test 'tab counts' do
